requirements.txt
requests>=2.28
pyautogui>=0.9.55
Pillow>=9.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
jinja2>=3.0.0
flet>=0.9.0
pystray>=0.21.0
pyinstaller>=5.10
